requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
astroid==2.11.3
azure-core==1.23.1
azure-storage-file-share==12.7.0
click-plugins==1.1.1
dill==0.3.4
isodate==0.6.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
msrest==0.6.21
platformdirs==2.5.2
pylint==2.13.7
requests-oauthlib==1.3.1
shodan==1.27.0
tomli==2.0.1
typing-extensions==4.2.0
wrapt==1.14.0
XlsxWriter==3.0.3
